+++ Windows/PowerShell/Modules/Tests/Modules/Bootstrap/Resolve-BootstrapSteps.Tests.ps1
@@ -35,7 +35,7 @@ Describe "Resolve-BootstrapSteps" {
 					"DotnetEf", "EnvironmentVariables", "CondaEnvironments", "Taskbar", "SymbolicLinks")) {
 				$states[$name] | Should -BeTrue -Because "step [$name] defaults to on"
 			}
-			foreach ($name in @("MicrosoftActivationScripts", "Win11Debloat", "DeveloperMode", "NuGetConfig", "UpgradeAll", "CoreAiRules", "LockedStartLayout")) {
+			foreach ($name in @("MicrosoftActivationScripts", "Win11Debloat", "DeveloperMode", "NuGetConfig", "UpgradeAll", "CoreAiRules", "AiSkills", "LockedStartLayout")) {
 				$states[$name] | Should -BeFalse -Because "step [$name] is opt-in"
 			}
 		}
@@ -53,7 +53,7 @@ Describe "Resolve-BootstrapSteps" {
 		It "Should list the steps in Bootstrap execution order" {
 			$states = Resolve-BootstrapSteps
 
-			@($states.Keys) -join "," | Should -Be "RenameMachine,MicrosoftActivationScripts,Win11Debloat,ExecutionPolicy,DeveloperMode,PowerPlan,PowerButtonActions,SystemTheme,Locale,DisplayLanguage,KeyboardLayouts,NerdFont,PowerShellModules,SpecialFolders,WSL,WinGetApps,ScoopApps,ChocolateyApps,UpgradeAll,DotnetEf,EnvironmentVariables,CondaEnvironments,NuGetConfig,Taskbar,SymbolicLinks,CoreAiRules,LockedStartLayout"
+			@($states.Keys) -join "," | Should -Be "RenameMachine,MicrosoftActivationScripts,Win11Debloat,ExecutionPolicy,DeveloperMode,PowerPlan,PowerButtonActions,SystemTheme,Locale,DisplayLanguage,KeyboardLayouts,NerdFont,PowerShellModules,SpecialFolders,WSL,WinGetApps,ScoopApps,ChocolateyApps,UpgradeAll,DotnetEf,EnvironmentVariables,CondaEnvironments,NuGetConfig,Taskbar,SymbolicLinks,CoreAiRules,AiSkills,LockedStartLayout"
 		}
 
 		It "Should return the defaults when Configuration itself is null" {